Opti 1: Int8 instead of Int (5.7s -> 5.2s)

+++ algo/random.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-func LaunchAlgoRandom(tab []int) {
+func LaunchAlgoRandom(tab []int8) {
 	utils.PrintTab(tab)
 	correct := checker.BuildCorrectResult(utils.Size)
 	for !checker.CheckResult(tab, correct) {
@@ -19,8 +19,8 @@ func LaunchAlgoRandom(tab []int) {
 	utils.PrintTab(tab)
 }
 
-func getRandomNumber(max int) int {
+func getRandomNumber(max int) int8 {
 	s1 := rand.NewSource(time.Now().UnixNano())
 	r1 := rand.New(s1)
-	return (r1.Intn(max))
+	return (int8(r1.Intn(max)))
 }
